+++ packages/insomnia/src/ui/components/modals/workspace-duplicate-modal.tsx
@@ -9,8 +9,8 @@ import { sortProjects } from '../../../models/helpers/project';
 import * as models from '../../../models/index';
 import type { Project } from '../../../models/project';
 import type { Workspace } from '../../../models/workspace';
+import { scopeToBgColorMap, scopeToIconMap, scopeToTextColorMap } from '../../routes/$organizationId.project';
 import { useOrganizationLoaderData } from '../../routes/organization';
-import { scopeToBgColorMap, scopeToIconMap, scopeToTextColorMap } from '../../routes/project';
 import { Modal, type ModalHandle, type ModalProps } from '../base/modal';
 import { ModalBody } from '../base/modal-body';
 import { ModalFooter } from '../base/modal-footer';
