+++ packages/insomnia/src/ui/components/project/project-settings-form.tsx
@@ -31,9 +31,14 @@ import {
   type Project,
 } from '../../../models/project';
 import type { StorageRules } from '../../organization-utils';
+import {
+  scopeToBgColorMap,
+  scopeToIconMap,
+  scopeToLabelMap,
+  scopeToTextColorMap,
+} from '../../routes/$organizationId.project';
 import type { UpdateProjectActionResult } from '../../routes/$organizationId.project.$projectId.update';
 import type { InitGitCloneResult } from '../../routes/git-project-actions';
-import { scopeToBgColorMap, scopeToIconMap, scopeToLabelMap, scopeToTextColorMap } from '../../routes/project';
 import { ErrorBoundary } from '../error-boundary';
 import { CustomRepositorySettingsFormGroup } from '../git-credentials/custom-repository-settings-form';
 import { GitHubRepositorySetupFormGroup } from '../git-credentials/github-repository-settings-form';
