+++ packages/insomnia/src/ui/components/settings/import-export.tsx
@@ -25,7 +25,6 @@ import { ImportModal } from 'insomnia/src/ui/components/modals/import-modal';
 import { SelectModal } from 'insomnia/src/ui/components/modals/select-modal';
 import type { WorkspaceLoaderData } from 'insomnia/src/ui/routes/$organizationId.project.$projectId.workspace';
 import { useOrganizationLoaderData } from 'insomnia/src/ui/routes/organization';
-import type { ListWorkspacesLoaderData } from 'insomnia/src/ui/routes/project';
 import { useRootLoaderData } from 'insomnia/src/ui/routes/root';
 import type { UntrackedProjectsLoaderData } from 'insomnia/src/ui/routes/untracked-projects';
 import React, { type FC, Fragment, useEffect, useState } from 'react';
@@ -33,6 +32,7 @@ import { Button, Heading, ListBox, ListBoxItem, Popover, Select, SelectValue } f
 import { useFetcher, useParams } from 'react-router';
 import { useRouteLoaderData } from 'react-router';
 
+import type { ListWorkspacesLoaderData } from '../../routes/$organizationId.project';
 import { AlertModal } from '../modals/alert-modal';
 
 const VALUE_YAML = 'yaml';
