+++ packages/insomnia/src/ui/routes/$organizationId.project.tsx
@@ -63,7 +63,6 @@ import {
 import { isDesign, scopeToActivity, type Workspace, type WorkspaceScope } from '../../models/workspace';
 import type { WorkspaceMeta } from '../../models/workspace-meta';
 import { VCSInstance } from '../../sync/vcs/insomnia-sync';
-import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
 import { getInitialRouteForOrganization } from '../../utils/router';
 import { SegmentEvent } from '../analytics';
@@ -86,6 +85,7 @@ import { TimeFromNow } from '../components/time-from-now';
 import { useInsomniaEventStreamContext } from '../context/app/insomnia-event-stream-context';
 import { useLoaderDeferData } from '../hooks/use-loader-defer-data';
 import { useOrganizationPermissions } from '../hooks/use-organization-features';
+import { insomniaFetch } from '../insomniaFetch';
 import { DEFAULT_STORAGE_RULES } from '../organization-utils';
 import { type OrganizationStorageLoaderData } from './$organizationId.storage-rules';
 import { useOrganizationLoaderData } from './organization';
